requirements.txt
numpy==1.23.4
pdf2image==1.17.0
Pillow==11.1.0
